requirements.txt
# Flask and WSGI server
Flask==3.0.0
gunicorn==21.2.0

# CORS support
flask-cors==4.0.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# Environment variables
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# HTTP testing
responses==0.24.1
